Latest Updates

+++ loginGUI.py
@@ -2,8 +2,9 @@ from tkinter import *
 from tkinter import messagebox
 from adminGUI import *
 from userGUI import *
-from encryption import *
 import json
+import string
+
 class Login_Gui(Tk):
     def __init__(self):
         super().__init__()
@@ -21,12 +22,10 @@ class Login_Gui(Tk):
         self.canvas.create_image(180, 160, image = self.logo)
         self.canvas.grid(column = 1, row = 0)
 
-
         self.username_label = Label(text = "Username:", bg = "light gray", highlightthickness = 0)
         self.username_label.grid(column = 1, row = 1)
         self.password_label = Label(text = "Password:", bg = "light gray", highlightthickness = 0)
         self.password_label.grid(column = 1, row = 3)
-
 
         self.login_logo = PhotoImage(file = "Assets/log-in.png")
         self.login_button = Button(image = self.login_logo, bd = 0, bg = "light gray", highlightthickness = 0, command = self.details_check)
@@ -35,12 +34,13 @@ class Login_Gui(Tk):
         self.signup_button = Button(image = self.signup_logo, bd = 0, bg = "light gray", highlightthickness = 0, command = self.save)
         self.signup_button.grid(column = 2, row = 5)
 
+
     def save(self):
         username = self.username_entry.get()
         password = self.password_entry.get()
         new_data = {username:
         {
-            "password" : password,
+            "password" : self.encode(password),
         }
     }
 
@@ -60,14 +60,18 @@ class Login_Gui(Tk):
                         json.dump(new_data, data_file, indent = 4)
             
                 else:
-                    data.update(new_data)
-                    with open("Login Details.json","w") as data_file:
-                        json.dump(data, data_file, indent = 4)
+                    if username in data:
+                        messagebox.showwarning(title = "Oops", message = "Username already exists")
+                    else:
+                        data.update(new_data)
+                        with open("Login Details.json","w") as data_file:
+                            json.dump(data, data_file, indent = 4)
             
                 finally:
                     self.username_entry.delete(0, END)
                     self.username_entry.focus()
                     self.password_entry.delete(0, END)
+
 
     def details_check(self):
         username = self.username_entry.get()
@@ -78,12 +82,31 @@ class Login_Gui(Tk):
         except FileNotFoundError:
             messagebox.showwarning(title = "Error", message = "No Data File Found")
         else:
-            if username in data and password == data[username]["password"]:
+            if username in data and self.encode(password) == data[username]["password"]:
                 self.loggedin_window()
-            elif username in data and password != data[username]["password"]:
+            elif username in data and self.encode(password) != data[username]["password"]:
                 messagebox.showwarning(title = "Error", message = "Password or username is incorrect")
             else:
                 messagebox.showwarning(title = "Error", message = "No details where found")
+
+
+    def encode(self, password):
+        alphabet_lower = string.ascii_lowercase
+        alphabet_upper = string.ascii_uppercase  
+        shift = 5
+        end_text = ""
+        for char in password:
+            if char in alphabet_lower:
+                position = alphabet_lower.index(char)
+                new_position = (position + shift) % 26
+                end_text += alphabet_lower[new_position]
+            elif char in alphabet_upper:
+                position = alphabet_upper.index(char)
+                new_position = (position + shift) % 26
+                end_text += alphabet_upper[new_position]
+            else:
+                end_text += char
+        return end_text
 
 
     def loggedin_window(self):
